pyproject.toml
[tool.poetry]
name = "kvdomingo-xyz"
version = "1.0.0"
description = ""
authors = ["Kenneth V. Domingo <hello@kvdomingo.xyz>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "4.0"
cloudinary = "^1.29.0"
dj-database-url = "^1.0.0"
django-admin-relation-links = "^0.2.5"
django-cors-headers = "^3.13.0"
django-filter = "^22.1"
django-ordered-model = "^3.6"
django-seo-js = "^0.3.5"
django-tinymce = "^3.5.0"
djangorestframework = "^3.13.1"
djangorestframework-camel-case = "^1.3.0"
gunicorn = {extras = ["gthread"], version = "^20.1.0"}
Jinja2 = "^3.1.2"
psycopg2 = "^2.9.3"
python-dotenv = "^0.20.0"
requests = "^2.28.1"
whitenoise = "^6.2.0"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py310']

[tool.isort]
profile = "black"
